+++ index.d.ts
@@ -4,18 +4,6 @@ type Point = GeoJSON.Point
 type LineString = GeoJSON.LineString
 type MultiLineString = GeoJSON.MultiLineString
 
-// /**
-//  * Finds the angle between 3 points.
-//  */
-// export function findAngle<T extends Point>(
-//     startPoint: T | GeoJSON.Feature<T> | number[],
-//     midPoint: T | GeoJSON.Feature<T> | number[],
-//     endPoint: T | GeoJSON.Feature<T> | number[],
-//     options?: {
-//         precision?: number
-//     }
-// ): number
-
 /**
  * Impossible Angle, this validator detects lines with less likely turning angles.
  */