+++ index.js
@@ -1,4 +1,2 @@
-module.exports = {
-  impossibleAngle: require('./validators/impossible-angle/index'),
-  closestEndNodes: require('./validators/closest-end-nodes/index')
-}
+export { default as impossibleAngle } from './validators/impossible-angle/index'
+export { default as closestEndNodes } from './validators/closest-end-nodes/index'