+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.3.0 - 2017-09-29
 
-- Convert back to ES5 module.exports
+- Improve ES module export/import
 
 ## 0.2.0 - 2017-09-26